Round case 4 cursor/mode dimensions through float
Apply the float round-trip (__cvt_fp2unsigned) to the special-mode
cursor and mode panel texture dimensions, matching the original.

DrawOptionMenu 52.12% -> 52.32%, unit 70.13% -> 70.25%.

+++ src/MenuUtil.cpp
@@ -1480,8 +1480,8 @@ void CMenuPcs::DrawOptionMenu()
 		for (int i = 0; i < 4; i++, y += 0x28, uvY += 0x20, modeU += 0x40) {
 			if ((m_specialModeEdit != 0) && (m_specialModeCursor == i)) {
 				CTexture* cursorPanel = GetTextureSetTexture(textureSet, 4);
-				unsigned int cursorWidth = cursorPanel->m_width;
-				unsigned int cursorHeight = cursorPanel->m_height;
+				unsigned int cursorWidth = static_cast<unsigned int>(static_cast<float>(cursorPanel->m_width));
+				unsigned int cursorHeight = static_cast<unsigned int>(static_cast<float>(cursorPanel->m_height));
 				gUtil.CalcUV(uv0.x, uv0.y, cursorWidth - 0x30, 0, cursorWidth, cursorHeight);
 				gUtil.CalcUV(uv1.x, uv1.y, cursorWidth, 0x28, cursorWidth, cursorHeight);
 				gUtil.RenderTextureQuad(326.0f, 128.0f + static_cast<float>(y), 48.0f, FLOAT_80333588,
@@ -1501,8 +1501,8 @@ void CMenuPcs::DrawOptionMenu()
 			}
 
 			CTexture* modePanel = GetTextureSetTexture(textureSet, 7);
-			unsigned int modeWidth = modePanel->m_width;
-			unsigned int modeHeight = modePanel->m_height;
+			unsigned int modeWidth = static_cast<unsigned int>(static_cast<float>(modePanel->m_width));
+			unsigned int modeHeight = static_cast<unsigned int>(static_cast<float>(modePanel->m_height));
 			gUtil.CalcUV(uv0.x, uv0.y, modeWidth - 0x30, uvY, modeWidth, modeHeight);
 			gUtil.CalcUV(uv1.x, uv1.y, modeWidth, uvY + 0x18, modeWidth, modeHeight);
 			const float modeX = static_cast<float>(
